enabling local profile dependency and code seperation
